Refactored methods, changing int to long to reflect changes made in different user models.

+++ qoajad_backend/src/main/java/com/qoajad/backend/utils/ValidationUtils.java
@@ -5,21 +5,21 @@ public class ValidationUtils {
     // This class cannot be instantiated.
     private ValidationUtils() {}
 
-    public static boolean requireEqual(final int left, final int right, final String message) {
+    public static boolean requireEqual(final long left, final long right, final String message) {
         if (left == right) {
             return true;
         }
         throw new IllegalStateException(message);
     }
 
-    public static boolean requireNotEqual(final int left, final int right, final String message) {
+    public static boolean requireNotEqual(final long left, final long right, final String message) {
         if (left != right) {
             return true;
         }
         throw new IllegalStateException(message);
     }
 
-    public static boolean requireLeftGreaterThanRight(final int left, final int right, final String message) {
+    public static boolean requireLeftGreaterThanRight(final long left, final long right, final String message) {
         if (left > right) {
             return true;
         }
